Fixed format and merge issues
